@@ -116,7 +118,10 @@
     "tslint": "~6.1.0",
     "ttypescript": "^1.5.12",
     "typedoc": "^0.16.4",
-    "typescript": ">=4.0.0 <4.1.0"
+    "typescript": ">=4.2.3 <4.3.0"
   },
-  "typings": "index.d.ts"
+  "typings": "index.d.ts",
+  "resolutions": {
+    "webpack": "5.4.0"
+  }
 }
