@@ -17,17 +17,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.0.0",
-    "@angular/common": "~11.0.0",
-    "@angular/compiler": "~11.0.0",
-    "@angular/core": "~11.0.0",
+    "@angular/animations": "~12.0.0",
+    "@angular/common": "~12.0.0",
+    "@angular/compiler": "~12.0.0",
+    "@angular/core": "~12.0.0",
     "@angular/fire": "../dist/packages-dist",
-    "@angular/forms": "~11.0.0",
-    "@angular/platform-browser": "~11.0.0",
-    "@angular/platform-browser-dynamic": "~11.0.0",
-    "@angular/platform-server": "~11.0.0",
-    "@angular/router": "~11.0.0",
-    "@angular/service-worker": "^11.0.0",
+    "@angular/forms": "~12.0.0",
+    "@angular/platform-browser": "~12.0.0",
+    "@angular/platform-browser-dynamic": "~12.0.0",
+    "@angular/platform-server": "~12.0.0",
+    "@angular/router": "~12.0.0",
+    "@angular/service-worker": "^12.0.0",
     "@nguniversal/express-engine": "~10.1.0",
     "core-js": "^3.6.5",
     "firebase": "^8.0.0",
@@ -36,14 +36,14 @@
     "rxjs": "~6.6.3",
     "tslib": "^2.0.1",
     "whatwg-fetch": "^3.4.1",
-    "zone.js": "0.10.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.1001.3",
-    "@angular-devkit/build-angular": "~0.1100.0",
-    "@angular/cli": "~11.0.0",
-    "@angular/compiler-cli": "~11.0.0",
-    "@angular/language-service": "~11.0.0",
+    "@angular-devkit/architect": "~0.1200.0",
+    "@angular-devkit/build-angular": "~12.0.0",
+    "@angular/cli": "~12.0.0",
+    "@angular/compiler-cli": "~12.0.0",
+    "@angular/language-service": "~12.0.0",
     "@firebase/app-types": "^0.6.1",
     "@nguniversal/builders": "^10.1.0",
     "@types/jasmine": "~3.5.0",
@@ -62,13 +62,15 @@
     "inquirer-autocomplete-prompt": "^1.0.1",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.1",
+    "jsonc-parser": "^3.0.0",
+    "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
+    "open": "^7.0.3",
     "ts-node": "~9.0.0",
     "tslint": "~6.1.3",
-    "typescript": "~4.0.5"
+    "typescript": "~4.2.4"
   }
 }
