@@ -22,7 +22,7 @@ export class FirestoreOfflineComponent implements OnInit {
 
   constructor(state: TransferState, firestore: AngularFirestoreOffline) {
     const doc = firestore.doc('test/1');
-    const key = makeStateKey(doc.ref.path);
+    const key = makeStateKey<unknown>(doc.ref.path);
     const existing = state.get(key, undefined);
     this.testDocValue$ = firestore.doc('test/1').valueChanges().pipe(
       trace('firestore'),
