@@ -21,7 +21,7 @@ export class FirestoreComponent implements OnInit {
 
   constructor(state: TransferState, firestore: AngularFirestore) {
     const doc = firestore.doc('test/1');
-    const key = makeStateKey(doc.ref.path);
+    const key = makeStateKey<unknown>(doc.ref.path);
     const existing = state.get(key, undefined);
     this.testDocValue$ = firestore.doc('test/1').valueChanges().pipe(
       trace('firestore'),
