@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { AngularFireModule, FirebaseApp } from '@angular/fire';
-import { AngularFireAnalytics, AngularFireAnalyticsModule } from './public_api';
+import { AngularFireAnalytics, AngularFireAnalyticsModule } from '@angular/fire/analytics';
 import { COMMON_CONFIG } from '../test-config';
 import { rando } from '../firestore/utils.spec';
 
@@ -22,7 +22,7 @@ describe('AngularFireAnalytics', () => {
   });
 
   afterEach(() => {
-    app.delete();
+    app.delete().catch();
   });
 
   it('should be exist', () => {
