@@ -20,7 +20,16 @@ export interface FirebaseTools {
   };
 
   logger: {
-    add(...args: any[]): any
+    // firebase-tools v8
+    add: (...args: any[]) => any
+    // firebase-tools v9
+    logger: {
+      add: (...args: any[]) => any;
+    }
+  };
+
+  cli: {
+    version(): string;
   };
 
   login(): Promise<void>;
