@@ -1,7 +1,9 @@
 import { readFileSync } from 'fs';
-import { FirebaseRc, Project, WorkspaceProject } from './interfaces';
+import { FirebaseRc, Project, Workspace, WorkspaceProject } from './interfaces';
 import { join } from 'path';
 import { isUniversalApp } from './ng-add-ssr';
+import { SchematicsException, Tree } from '@angular-devkit/schematics';
+import { DeployOptions } from './ng-add-common';
 
 export async function listProjects() {
   const firebase = require('firebase-tools');
@@ -41,6 +43,60 @@ const searchProjects = (projects: Project[]) => {
   };
 };
 
+
+export function getWorkspace(
+  host: Tree
+): { path: string; workspace: Workspace } {
+  const possibleFiles = ['/angular.json', '/.angular.json'];
+  const path = possibleFiles.filter(p => host.exists(p))[0];
+
+  const configBuffer = host.read(path);
+  if (configBuffer === null) {
+    throw new SchematicsException(`Could not find angular.json`);
+  }
+
+  // We can not depend on this library to have be included in older (or newer) Angular versions.
+  // Require here, since the schematic will add it to the package.json and install it before
+  // continuing.
+  const { parse }: typeof import('jsonc-parser') = require('jsonc-parser');
+
+  const workspace = parse(configBuffer.toString()) as Workspace|undefined;
+  if (!workspace) {
+    throw new SchematicsException('Could not parse angular.json');
+  }
+
+  return {
+    path,
+    workspace
+  };
+}
+
+export const getProject = (options: DeployOptions, host: Tree) => {
+  const { workspace } = getWorkspace(host);
+  const projectName = options.project || workspace.defaultProject;
+
+  if (!projectName) {
+    throw new SchematicsException(
+      'No Angular project selected and no default project in the workspace'
+    );
+  }
+
+  const project = workspace.projects[projectName];
+  if (!project) {
+    throw new SchematicsException(
+      'The specified Angular project is not defined in this workspace'
+    );
+  }
+
+  if (project.projectType !== 'application') {
+    throw new SchematicsException(
+      `Deploy requires an Angular project type of "application" in angular.json`
+    );
+  }
+
+  return {project, projectName};
+};
+
 export const projectPrompt = (projects: Project[]) => {
   const inquirer = require('inquirer');
   inquirer.registerPrompt(
@@ -55,7 +111,7 @@ export const projectPrompt = (projects: Project[]) => {
   });
 };
 
-export const projectTypePrompt = (project: WorkspaceProject) => {
+export const projectTypePrompt = (project: WorkspaceProject): Promise<{ universalProject: boolean }> => {
   if (isUniversalApp(project)) {
     return require('inquirer').prompt({
       type: 'confirm',
