@@ -1,13 +1,13 @@
-import { forkJoin, from } from 'rxjs';
+import { forkJoin } from 'rxjs';
 import { mergeMap, tap } from 'rxjs/operators';
 import { TestBed } from '@angular/core/testing';
 import { AngularFireModule, FIREBASE_APP_NAME, FIREBASE_OPTIONS, FirebaseApp } from '@angular/fire';
-import { AngularFireStorage, AngularFireStorageModule, AngularFireUploadTask, BUCKET } from './public_api';
+import { AngularFireStorage, AngularFireStorageModule, AngularFireUploadTask, BUCKET, fromTask } from '@angular/fire/storage';
 import { COMMON_CONFIG } from '../test-config';
 import { rando } from '../firestore/utils.spec';
-import { GetDownloadURLPipe } from './pipes/storageUrl.pipe';
 import { ChangeDetectorRef } from '@angular/core';
 import 'firebase/storage';
+import firebase from 'firebase/app';
 
 if (typeof XMLHttpRequest === 'undefined') {
   globalThis.XMLHttpRequest = require('xhr2');
@@ -43,7 +43,7 @@ describe('AngularFireStorage', () => {
   });
 
   afterEach(() => {
-    app.delete();
+    app.delete().catch();
   });
 
   it('should exist', () => {
@@ -63,15 +63,21 @@ describe('AngularFireStorage', () => {
     it('should upload and delete a file', (done) => {
       const data = { angular: 'fire' };
       const blob = blobOrBuffer(JSON.stringify(data), { type: 'application/json' });
-      const ref = afStorage.ref('af.json');
+      const ref = afStorage.ref(rando());
       const task = ref.put(blob);
+      let emissionCount = 0;
+      let lastSnap: firebase.storage.UploadTaskSnapshot;
       task.snapshotChanges()
         .subscribe(
           snap => {
+            lastSnap = snap;
+            emissionCount++;
             expect(snap).toBeDefined();
           },
           done.fail,
           () => {
+            expect(lastSnap.state).toBe('success');
+            expect(emissionCount).toBeGreaterThan(0);
             ref.delete().subscribe(done, done.fail);
           });
     });
@@ -79,7 +85,7 @@ describe('AngularFireStorage', () => {
     it('should upload a file and observe the download url', (done) => {
       const data = { angular: 'fire' };
       const blob = blobOrBuffer(JSON.stringify(data), { type: 'application/json' });
-      const ref = afStorage.ref('af.json');
+      const ref = afStorage.ref(rando());
       ref.put(blob).then(() => {
         const url$ = ref.getDownloadURL();
         url$.subscribe(
@@ -97,12 +103,74 @@ describe('AngularFireStorage', () => {
     it('should resolve the task as a promise', (done) => {
       const data = { angular: 'promise' };
       const blob = blobOrBuffer(JSON.stringify(data), { type: 'application/json' });
-      const ref = afStorage.ref('af.json');
+      const ref = afStorage.ref(rando());
       const task: AngularFireUploadTask = ref.put(blob);
       task.then(snap => {
         expect(snap).toBeDefined();
         done();
       }).catch(done.fail);
+    });
+
+    it('should cancel the task', (done) => {
+      const data = { angular: 'promise' };
+      const blob = blobOrBuffer(JSON.stringify(data), { type: 'application/json' });
+      const ref = afStorage.ref(rando());
+      const task: AngularFireUploadTask = ref.put(blob);
+      let emissionCount = 0;
+      let lastSnap: firebase.storage.UploadTaskSnapshot;
+      task.snapshotChanges().subscribe(snap => {
+        emissionCount++;
+        lastSnap = snap;
+        if (emissionCount === 1) {
+          task.cancel();
+        }
+      }, () => {
+        // TODO investigate, this doesn't appear to work...
+        // https://github.com/firebase/firebase-js-sdk/issues/4158
+        // expect(lastSnap.state).toEqual('canceled');
+        done();
+      }, done.fail);
+    });
+
+    it('should be able to pause/resume the task', (done) => {
+      const data = { angular: 'promise' };
+      const blob = blobOrBuffer(JSON.stringify(data), { type: 'application/json' });
+      const ref = afStorage.ref(rando());
+      const task: AngularFireUploadTask = ref.put(blob);
+      let paused = false;
+      task.pause();
+      task.snapshotChanges().subscribe(snap => {
+        if (snap.state === 'paused') {
+          paused = true;
+          task.resume();
+        }
+      }, done.fail, () => {
+        expect(paused).toBeTruthy();
+        done();
+      });
+    });
+
+    it('should work with an already finished task', (done) => {
+      const data = { angular: 'promise' };
+      const blob = blobOrBuffer(JSON.stringify(data), { type: 'application/json' });
+      const ref = afStorage.storage.ref(rando());
+      const task = ref.put(blob);
+      let emissionCount = 0;
+      let lastSnap: firebase.storage.UploadTaskSnapshot;
+      task.then(_snap => {
+        fromTask(task).subscribe(
+            snap => {
+              lastSnap = snap;
+              emissionCount++;
+              expect(snap).toBeDefined();
+            },
+            done.fail,
+            () => {
+              expect(lastSnap.state).toBe('success');
+              expect(emissionCount).toBe(1);
+              ref.delete().then(done, done.fail);
+            });
+      });
     });
 
   });
@@ -112,7 +180,7 @@ describe('AngularFireStorage', () => {
     it('it should upload, download, and delete', (done) => {
       const data = { angular: 'fire' };
       const blob = blobOrBuffer(JSON.stringify(data), { type: 'application/json' });
-      const ref = afStorage.ref('af.json');
+      const ref = afStorage.ref(rando());
       const task = ref.put(blob);
       // Wait for the upload
       forkJoin([task.snapshotChanges()])
@@ -131,7 +199,7 @@ describe('AngularFireStorage', () => {
     it('should upload, get metadata, and delete', (done) => {
       const data = { angular: 'fire' };
       const blob = blobOrBuffer(JSON.stringify(data), { type: 'application/json' });
-      const ref = afStorage.ref('af.json');
+      const ref = afStorage.ref(rando());
       const task = ref.put(blob, { customMetadata: { blah: 'blah' } });
       // Wait for the upload
       forkJoin([task.snapshotChanges()])
@@ -177,7 +245,7 @@ describe('AngularFireStorage w/options', () => {
   });
 
   afterEach(() => {
-    app.delete();
+    app.delete().catch();
   });
 
   describe('<constructor>', () => {
@@ -208,7 +276,8 @@ describe('AngularFireStorage w/options', () => {
       it('it should upload, download, and delete', (done) => {
         const data = { angular: 'fire' };
         const blob = blobOrBuffer(JSON.stringify(data), { type: 'application/json' });
-        const ref = afStorage.ref('af.json');
+        const name = rando();
+        const ref = afStorage.ref(name);
         const task = ref.put(blob);
         // Wait for the upload
         forkJoin([task.snapshotChanges()])
@@ -216,7 +285,7 @@ describe('AngularFireStorage w/options', () => {
             // get the url download
             mergeMap(() => ref.getDownloadURL()),
             // assert the URL
-            tap(url => expect(url).toMatch(new RegExp(`https:\\/\\/firebasestorage\\.googleapis\\.com\\/v0\\/b\\/${storageBucket}\\/o\\/af\\.json`))),
+            tap(url => expect(url).toMatch(new RegExp(`https:\\/\\/firebasestorage\\.googleapis\\.com\\/v0\\/b\\/${storageBucket}\\/o\\/${name}`))),
             // Delete the file
             mergeMap(() => ref.delete())
           )
