@@ -1,5 +1,20 @@
 Jasmine = require('jasmine');
 
+require('reflect-metadata');
+require('zone.js/dist/zone-node');
+require('zone.js/dist/zone-testing');
+
+const { getTestBed } = require('@angular/core/testing');
+const { platformServerTesting, ServerTestingModule } = require('@angular/platform-server/testing');
+
+global['globalThis'] = require('globalthis/polyfill')();
+
+// First, initialize the Angular testing environment.
+getTestBed().initTestEnvironment(
+    ServerTestingModule,
+    platformServerTesting()
+);
+
 jasmine = new Jasmine();
 jasmine.loadConfig({
     spec_dir: '.',
@@ -8,23 +23,5 @@ jasmine.loadConfig({
         'dist/out-tsc/jasmine/**/*.spec.js'
     ]
 });
-
-require('reflect-metadata');
-require('zone.js/dist/zone-node');
-require('zone.js/dist/async-test');
-require('zone.js/dist/sync-test');
-require('zone.js/dist/fake-async-test');
-require('zone.js/dist/proxy');
-require('zone.js/dist/zone-patch-rxjs');
-require('zone.js/dist/task-tracking');
-
-const { getTestBed } = require('@angular/core/testing');
-const { platformServerTesting, ServerTestingModule } = require('@angular/platform-server/testing');
-
-// First, initialize the Angular testing environment.
-getTestBed().initTestEnvironment(
-    ServerTestingModule,
-    platformServerTesting()
-);
 
 jasmine.execute();
